pyproject.toml
[project]
name = "nrct-certificate-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
]
